chore: Cleanup Slack/Mattermost/MSTeams notifiers
These will need some refactoring for #8840 and this will hopefully
simplify this.
