+++ packages/server/graphql/mutations/helpers/notifications/MattermostNotifier.ts
@@ -21,8 +21,7 @@ import {
   makeHackedFieldButtonValue
 } from './makeMattermostAttachments'
 import {NotificationIntegrationHelper} from './NotificationIntegrationHelper'
-import {Notifier} from './Notifier'
-import {getTeamPromptResponsesByMeetingId} from '../../../../postgres/queries/getTeamPromptResponsesByMeetingIds'
+import {createNotifier} from './Notifier'
 import {analytics} from '../../../../utils/analytics/analytics'
 
 const notifyMattermost = async (
@@ -322,90 +321,13 @@ async function getMattermost(dataLoader: DataLoaderWorker, teamId: string, userI
     .get('bestTeamIntegrationProviders')
     .load({service: 'mattermost', teamId, userId})
   return provider
-    ? MattermostNotificationHelper({
-        ...(provider as IntegrationProviderMattermost),
-        userId
-      })
-    : null
+    ? [
+        MattermostNotificationHelper({
+          ...(provider as IntegrationProviderMattermost),
+          userId
+        })
+      ]
+    : []
 }
 
-async function loadMeetingTeam(dataLoader: DataLoaderWorker, meetingId: string, teamId: string) {
-  const [team, meeting] = await Promise.all([
-    dataLoader.get('teams').load(teamId),
-    dataLoader.get('newMeetings').load(meetingId)
-  ])
-  return {
-    meeting,
-    team
-  }
-}
-
-export const MattermostNotifier: Notifier = {
-  async startMeeting(dataLoader: DataLoaderWorker, meetingId: string, teamId: string) {
-    const {meeting, team} = await loadMeetingTeam(dataLoader, meetingId, teamId)
-    if (!meeting || !team) return
-    ;(await getMattermost(dataLoader, team.id, meeting.facilitatorUserId))?.startMeeting(
-      meeting,
-      team
-    )
-  },
-
-  async endMeeting(dataLoader: DataLoaderWorker, meetingId: string, teamId: string) {
-    const {meeting, team} = await loadMeetingTeam(dataLoader, meetingId, teamId)
-    if (!meeting || !team) return
-    ;(await getMattermost(dataLoader, team.id, meeting.facilitatorUserId))?.endMeeting(
-      meeting,
-      team,
-      null
-    )
-  },
-
-  async startTimeLimit(
-    dataLoader: DataLoaderWorker,
-    scheduledEndTime: Date,
-    meetingId: string,
-    teamId: string
-  ) {
-    const {meeting, team} = await loadMeetingTeam(dataLoader, meetingId, teamId)
-    if (!meeting || !team) return
-    ;(await getMattermost(dataLoader, team.id, meeting.facilitatorUserId))?.startTimeLimit(
-      scheduledEndTime,
-      meeting,
-      team
-    )
-  },
-
-  async endTimeLimit(dataLoader: DataLoaderWorker, meetingId: string, teamId: string) {
-    const {meeting, team} = await loadMeetingTeam(dataLoader, meetingId, teamId)
-    if (!meeting || !team) return
-    ;(await getMattermost(dataLoader, team.id, meeting.facilitatorUserId))?.endTimeLimit(
-      meeting,
-      team
-    )
-  },
-
-  async integrationUpdated(dataLoader: DataLoaderWorker, teamId: string, userId: string) {
-    ;(await getMattermost(dataLoader, teamId, userId))?.integrationUpdated()
-  },
-
-  async standupResponseSubmitted(
-    dataLoader: DataLoaderWorker,
-    meetingId: string,
-    teamId: string,
-    userId: string
-  ) {
-    const [{meeting, team}, user, responses] = await Promise.all([
-      loadMeetingTeam(dataLoader, meetingId, teamId),
-      dataLoader.get('users').load(userId),
-      getTeamPromptResponsesByMeetingId(meetingId)
-    ])
-    const response = responses.find(({userId: responseUserId}) => responseUserId === userId)
-    if (!meeting || !team || !response || !user) return
-    ;(await getMattermost(dataLoader, teamId, userId))?.standupResponseSubmitted(
-      meeting,
-      team,
-      user,
-      response
-    )
-  }
-}
+export const MattermostNotifier = createNotifier(getMattermost)
