+++ packages/server/graphql/mutations/helpers/notifications/NotificationIntegrationHelper.ts
@@ -28,6 +28,8 @@ export type NotificationIntegration = {
     user: User,
     response: TeamPromptResponse
   ): Promise<NotifyResponse>
+
+  sendNotificationToUser?(notificationId: string, userId: string): Promise<NotifyResponse>
 }
 
 export type NotificationIntegrationHelper<NotificationChannel> = (
