+++ packages/server/graphql/mutations/helpers/notifications/SlackNotifier.ts
@@ -17,7 +17,7 @@ import {DataLoaderWorker} from '../../../graphql'
 import getSummaryText from './getSummaryText'
 import {makeButtons, makeHeader, makeSection, makeSections} from './makeSlackBlocks'
 import {NotificationIntegrationHelper} from './NotificationIntegrationHelper'
-import {Notifier} from './Notifier'
+import {createNotifier} from './Notifier'
 import SlackAuth from '../../../../database/types/SlackAuth'
 import {getTeamPromptResponsesByMeetingId} from '../../../../postgres/queries/getTeamPromptResponsesByMeetingIds'
 import {ErrorResponse, PostMessageResponse} from '../../../../../client/utils/SlackManager'
@@ -456,24 +456,14 @@ export const SlackSingleChannelNotifier: NotificationIntegrationHelper<SlackNoti
 
 async function getSlack(
   dataLoader: DataLoaderWorker,
-  event: SlackNotificationEvent,
-  teamId: string
+  teamId: string,
+  _userId: string,
+  event: SlackNotificationEvent
 ) {
   const notifications = await dataLoader
     .get('slackNotificationsByTeamIdAndEvent')
     .load({event, teamId})
   return notifications.map(SlackSingleChannelNotifier)
-}
-
-async function loadMeetingTeam(dataLoader: DataLoaderWorker, meetingId: string, teamId: string) {
-  const [team, meeting] = await Promise.all([
-    dataLoader.get('teams').load(teamId),
-    dataLoader.get('newMeetings').load(meetingId)
-  ])
-  return {
-    meeting,
-    team
-  }
 }
 
 const getDmSlackForMeeting = async (
@@ -510,84 +500,8 @@ const getDmSlackForMeeting = async (
   return userSlackAuths[0]
 }
 
-export const SlackNotifier: Notifier = {
-  async startMeeting(dataLoader: DataLoaderWorker, meetingId: string, teamId: string) {
-    const {meeting, team} = await loadMeetingTeam(dataLoader, meetingId, teamId)
-    if (!meeting || !team) return
-    const notifiers = await getSlack(dataLoader, 'meetingStart', team.id)
-    notifiers.forEach((notifier) => notifier.startMeeting(meeting, team))
-  },
-
-  async updateMeeting(dataLoader: DataLoaderWorker, meetingId: string, teamId: string) {
-    const {meeting, team} = await loadMeetingTeam(dataLoader, meetingId, teamId)
-    if (!meeting || !team) return
-    const notifiers = await getSlack(dataLoader, 'meetingStart', team.id)
-    notifiers.forEach((notifier) => notifier.updateMeeting?.(meeting, team))
-  },
-
-  async endMeeting(dataLoader: DataLoaderWorker, meetingId: string, teamId: string) {
-    const {meeting, team} = await loadMeetingTeam(dataLoader, meetingId, teamId)
-    const meetingResponses = await getTeamPromptResponsesByMeetingId(meetingId)
-    // const standupResponses: Array<{user: User; response: TeamPromptResponse}> = []
-    const standupResponses = await Promise.all(
-      meetingResponses.map(async (response) => {
-        const user = await dataLoader.get('users').loadNonNull(response.userId)
-        return {
-          user,
-          response
-        }
-      })
-    )
-    if (!meeting || !team) return
-    const notifiers = await getSlack(dataLoader, 'meetingEnd', team.id)
-    notifiers.forEach((notifier) => notifier.endMeeting(meeting, team, standupResponses))
-  },
-
-  async startTimeLimit(
-    dataLoader: DataLoaderWorker,
-    scheduledEndTime: Date,
-    meetingId: string,
-    teamId: string
-  ) {
-    const {meeting, team} = await loadMeetingTeam(dataLoader, meetingId, teamId)
-    if (!meeting || !team) return
-    const notifiers = await getSlack(dataLoader, 'MEETING_STAGE_TIME_LIMIT_START', team.id)
-    notifiers.forEach((notifier) => notifier.startTimeLimit(scheduledEndTime, meeting, team))
-  },
-
-  async endTimeLimit(dataLoader: DataLoaderWorker, meetingId: string, teamId: string) {
-    const {meeting, team} = await loadMeetingTeam(dataLoader, meetingId, teamId)
-    if (!meeting || !team) return
-    const notifiers = await getSlack(dataLoader, 'MEETING_STAGE_TIME_LIMIT_END', team.id)
-    notifiers.forEach((notifier) => notifier.endTimeLimit(meeting, team))
-  },
-
-  async integrationUpdated() {
-    // Slack sends a system message on its own
-  },
-
-  async standupResponseSubmitted(
-    dataLoader: DataLoaderWorker,
-    meetingId: string,
-    teamId: string,
-    userId: string
-  ) {
-    const [{meeting, team}, user, responses] = await Promise.all([
-      loadMeetingTeam(dataLoader, meetingId, teamId),
-      dataLoader.get('users').load(userId),
-      getTeamPromptResponsesByMeetingId(meetingId)
-    ])
-    const response = responses.find(({userId: responseUserId}) => responseUserId === userId)
-    if (!meeting || !team || !response || !user) {
-      return
-    }
-
-    const notifiers = await getSlack(dataLoader, 'STANDUP_RESPONSE_SUBMITTED', team.id)
-    notifiers.forEach((notifier) =>
-      notifier.standupResponseSubmitted(meeting, team, user, response)
-    )
-  },
-
+export const SlackNotifier = {
+  ...createNotifier(getSlack),
   async shareTopic(
     dataLoader: DataLoaderWorker,
     userId: string,
